# openapi/spec.yaml
openapi: 3.0.0
info:
  title: Products
  version: 1.0.0
paths:
...
  /customer/{id}:
    get:
      responses:
        "200":
          description: "update user"
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/Customers"
    patch:
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: "More Updates"
        required: true
        content:
          application/json:
            schema:
              type: object
              $ref: "#/components/schemas/Customers"
      responses:
        "200":
          description: "Updated user"
...
components:
  schemas:
...
    Customers:
      type: object
      properties:
        id:
          type: string
        created_at:
          type: string
        displayName:
          type: string
        cart:
          type: array
          items: string
        orders:
          type: array
          items: string
        email:
          type: string
...